IMO: gateway以下を細かくパッケージわけなくてもよさそう
