+++ gateways/git_api.go
@@ -1,4 +1,4 @@
-package gitapi
+package gateways
 
 import (
 	"context"
@@ -10,6 +10,14 @@ import (
 
 	"github.com/cloudnativedaysjp/reviewapp-operator/models"
 )
+
+type GitApiIFace interface {
+	WithCredential(username, token string) error
+	ListOpenPullRequests(ctx context.Context, org, repo string) ([]*models.PullRequest, error)
+	GetOpenPullRequest(ctx context.Context, org, repo string, prNum int) (*models.PullRequest, error)
+	CommentToPullRequest(ctx context.Context, pr models.PullRequest, comment string) error
+	GetCommitHashes(ctx context.Context, pr models.PullRequest) ([]string, error)
+}
 
 type GitApiDriver struct {
 	logger logr.Logger
