+++ wire/wire_gen.go
@@ -6,8 +6,7 @@
 package wire
 
 import (
-	"github.com/cloudnativedaysjp/reviewapp-operator/gateways/gitapi"
-	"github.com/cloudnativedaysjp/reviewapp-operator/gateways/gitcommand"
+	"github.com/cloudnativedaysjp/reviewapp-operator/gateways"
 	"github.com/cloudnativedaysjp/reviewapp-operator/services"
 	"github.com/go-logr/logr"
 	"k8s.io/utils/exec"
@@ -16,13 +15,13 @@ import (
 // Injectors from wire.go:
 
 func NewGitRemoteRepoAppService(l logr.Logger) (*services.GitRemoteRepoAppService, error) {
-	gitApiDriver := gitapi.NewGitApiDriver(l)
+	gitApiDriver := gateways.NewGitApiDriver(l)
 	gitRemoteRepoAppService := services.NewGitRemoteRepoAppService(gitApiDriver, l)
 	return gitRemoteRepoAppService, nil
 }
 
 func NewGitRemoteRepoInfraService(l logr.Logger, e exec.Interface) (*services.GitRemoteRepoInfraService, error) {
-	gitCommandDriver, err := gitcommand.NewGitCommandDriver(l, e)
+	gitCommandDriver, err := gateways.NewGitCommandDriver(l, e)
 	if err != nil {
 		return nil, err
 	}
